refactor: move code to plugin
