+++ packages/core/client/src/schema-initializer/items/index.tsx
@@ -47,6 +47,5 @@ export * from './TableCollectionFieldInitializer';
 export * from './TableSelectorInitializer';
 export * from './UpdateActionInitializer';
 export * from './UpdateSubmitActionInitializer';
-export * from './UploadActionInitializer';
 export * from './ViewActionInitializer';
