+++ packages/plugins/file-manager/src/client/index.tsx
@@ -1,8 +1,18 @@
-import { PluginManagerContext, SettingsCenterProvider, registerTemplate } from '@nocobase/client';
+import {
+  PluginManagerContext,
+  SchemaComponentOptions,
+  SchemaInitializerContext,
+  SchemaInitializerProvider,
+  SettingsCenterProvider,
+  registerTemplate,
+  useCollection,
+} from '@nocobase/client';
 import { forEach } from '@nocobase/utils/client';
 import React, { useContext } from 'react';
 import { FileStoragePane } from './FileStorage';
 import { FileStorageShortcut } from './FileStorageShortcut';
+import * as hooks from './hooks';
+import * as initializers from './initializers';
 import * as templates from './templates';
 
 // 注册之后就可以在 Crete collection 按钮中选择创建了
@@ -11,6 +21,28 @@ forEach(templates, (template, key: string) => {
 });
 
 export default function (props) {
+  const initializes = useContext(SchemaInitializerContext);
+  const hasUploadAction = initializes.TableActionInitializers.items[0].children.some(
+    (initialize) => initialize.component === 'UploadActionInitializer',
+  );
+  !hasUploadAction &&
+    initializes.TableActionInitializers.items[0].children.push({
+      type: 'item',
+      title: "{{t('Upload')}}",
+      component: 'UploadActionInitializer',
+      schema: {
+        'x-align': 'right',
+        'x-decorator': 'ACLActionProvider',
+        'x-acl-action-props': {
+          skipScopeCheck: true,
+        },
+      },
+      visible: () => {
+        const collection = useCollection();
+        return collection.template === 'file';
+      },
+    });
+
   const ctx = useContext(PluginManagerContext);
   return (
     <SettingsCenterProvider
@@ -35,7 +67,9 @@ export default function (props) {
           },
         }}
       >
-        {props.children}
+        <SchemaComponentOptions scope={hooks}>
+          <SchemaInitializerProvider components={initializers}>{props.children}</SchemaInitializerProvider>
+        </SchemaComponentOptions>
       </PluginManagerContext.Provider>
     </SettingsCenterProvider>
   );
