+++ packages/plugins/file-manager/src/client/initializers/index.ts
@@ -1,0 +1,1 @@
+export * from './UploadActionInitializer';
